[프로그래머스] 멀쩡한 사각형 오타수정

+++ Programmers/멀쩡한사각형_v2.cpp
@@ -30,7 +30,7 @@ long long solution(int w,int h)
     
     not_use *= g;
     
-    answer = w*h - not_use;
+    answer = (long long)w* (long long)h - not_use;
